Footer.js
The footer was a NIGHTMARE. For some reason it would appear in the middle of the page. I then got it to stick to the bottom with position: fixed but on the portfolio page (which had the most content) it would appear halfway down the scroll. I eventually found a youtube react css tutorial that used both the app.css and index.css to fix the issue and appear correctly on every page.