Portfolio.js
I initially had the portfolio set into two separate files. One was supposed to hold all the data for each project and the main portfolio page was supposed to use props to pull the data. This was to make sure I did not have to separately  created each card and style them. However I could not get it to work so I followed react website doucmentation on how to create a card and used that for each required project.