index.css
CSS is forever be the hardest part of any front end. I managed to use Vanilla CSS to style everything and used the recommended COLORs website to choose easy on the eyes color scheme. I also managed to whip together a media query that allows the app to appear correctly on smaller screens.